+++ src/HorusStudio.Maui.MaterialDesignControls/Builder/MaterialSnackbarOptions.cs
@@ -13,6 +13,7 @@ public sealed class MaterialSnackbarOptions
     public double? DefaultFontSize { get; set; }
     public Color? DefaultActionColor { get; set; }
     public double? DefaultActionSize { get; set; }
+    public TextDecorations? DefaultActionTextDecorations { get; set; }
     public Color? DefaultIconColor { get; set; }
     public TimeSpan? DefaultDuration { get; set; }
     public int? DefaultSpacing { get; set; }