+++ ReleaseNotes.md
@@ -2,6 +2,7 @@
 
 ## Content table <!-- omit from toc -->
 - [Versions](#versions)
+  - [1.0.0-preview.6](#100-preview6)
   - [1.0.0-preview.5](#100-preview5)
   - [1.0.0-preview.4](#100-preview4)
   - [1.0.0-preview.3](#100-preview3)
@@ -9,6 +10,11 @@
   - [1.0.0-preview.1](#100-preview1)
 
 ## Versions
+### [1.0.0-preview.6](https://www.nuget.org/packages/HorusStudio.Maui.MaterialDesignControls/1.0.0-preview.6)
+* Add LineBreakMode to inputs placeholder
+* Add TextDecorations to Snackbar action
+* [iOS] Bug: Fix Snackbar background color
+  
 ### [1.0.0-preview.5](https://www.nuget.org/packages/HorusStudio.Maui.MaterialDesignControls/1.0.0-preview.5)
 * Bug: Outlined inputs not rendered when overriding global style
 * Bug: Tint color fixed on MaterialIconButton